requirements.txt
pandas
aiohttp
python-binance